separate component - tweet/tweets container
